+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.22.0 (DEV)
 
-- ...
+- Add: Support for SO Green Disk and Screw custom tiles replacement.
 
 ## 0.21.0 (2024-08-23)